refactor(tools): extract a base http tool
